Fix cleanup for remotes
I ran into a problem where snapshots were not cleaned as only remote snapshots where pulled from a remote host and those were a bit too old to fit into the today check. To circumvent that i extended the check to include yesterday - the cleanup is handled by the schema anyway so it's just the condition to trigger the cleanup that is affected. That way i get my snapshots cleaned up and still the manager does not clean snapshots of no longer active replications which, i think, was the reason for this check in the first place.